Add workaround for Vault Agent UNIX Sockets
Vault Agent gives us the capability to listen on a UNIX socket for
connections. This is super handy for environments relying on UNIX
permissions for application access, whereas a TCP listener on localhost
doesn't give us the same kind of control. Support for this isn't baked
directly into HVAC, but there is a mechanism for clients to pass in a
requests-compatible interface. In lieu of direct support in HVAC, this
commit implements a special construction in this collection for handling
the UNIX Socket case. (Basically, if 'unix://' in vault_addr, then
handle the "special case" until HVAC adds direct support.)