Advise against strings in hashivault_pki_role params
This module, like Vault API, accepts both lists and strings with
comma-separated values for some parameters. However, the API always
returns lists. As such, the module isn't idempotent when strings are
used because it detects data changes, although these are only type and
format changes (ie. data are sementically the same).

This change adds a note about that in pki_role module doc.

Signed-off-by: Albin Kerouanton <albinker@gmail.com>