add a hashivault_ssh_role_list module
This module lists roles configured in a specified ssh secrets engine.

This module supports check mode (even though it does not perform any
write operations) to ensure a whole playbook can run in check mode and
proper data is still returned from this module for further tasks to use.

The code in this commit has been formatted with `black`.